+++ net/tests/SEALContextTests.cs
@@ -140,7 +140,7 @@ namespace SEALNetTest
             };
             EncryptionParameters parms = new EncryptionParameters(SchemeType.CKKS)
             {
-                PolyModulusDegree = 2 * slotSize,
+                PolyModulusDegree = 2 * (ulong)slotSize,
                 CoeffModulus = coeffModulus
             };
             SEALContext context = SEALContext.Create(parms);
